Extract HtmlTableFormatter
